+++ Thunderstore/CHANGELOG.md
@@ -2,6 +2,8 @@
 - Added death screenshot `configurable On/Off`
 - Added config for death webhook URL, to separate from notifications
 - If you want to keep using notification URL, just input same URL
+- new in-game chat command: `/selfie`
+- Tweaked config layout, added category `Death Feed`
 
 # 1.1.1
 - **Fixed**: Resolved TaskCanceledException during WebSocket reconnection attempts